Refactor and clarify the block parser code
