+++ packages/blocks/src/api/index.js
@@ -33,11 +33,11 @@ export {
 // block. For composition, it also means inner blocks can effectively be child
 // components whose mechanisms can be shielded from the `edit` implementation
 // and just passed along.
+export { default as parse } from './parser';
 export {
-	default as parse,
 	getBlockAttributes,
 	parseWithAttributeSchema,
-} from './parser';
+} from './parser/get-block-attributes';
 
 // While block transformations account for a specific surface of the API, there
 // are also raw transformations which handle arbitrary sources not made out of
