+++ packages/blocks/src/api/parser/convert-legacy-block.js
@@ -1,0 +1,59 @@
+/**
+ * Convert legacy blocks to their canonical form. This function is used
+ * both in the parser level for previous content and to convert such blocks
+ * used in Custom Post Types templates.
+ *
+ * @param {string} name       The block's name
+ * @param {Object} attributes The block's attributes
+ *
+ * @return {[string, Object]} The block's name and attributes, changed accordingly if a match was found
+ */
+export function convertLegacyBlockNameAndAttributes( name, attributes ) {
+	const newAttributes = { ...attributes };
+	// Convert 'core/cover-image' block in existing content to 'core/cover'.
+	if ( 'core/cover-image' === name ) {
+		name = 'core/cover';
+	}
+
+	// Convert 'core/text' blocks in existing content to 'core/paragraph'.
+	if ( 'core/text' === name || 'core/cover-text' === name ) {
+		name = 'core/paragraph';
+	}
+
+	// Convert derivative blocks such as 'core/social-link-wordpress' to the
+	// canonical form 'core/social-link'.
+	if ( name && name.indexOf( 'core/social-link-' ) === 0 ) {
+		// Capture `social-link-wordpress` into `{"service":"wordpress"}`
+		newAttributes.service = name.substring( 17 );
+		name = 'core/social-link';
+	}
+
+	// Convert derivative blocks such as 'core-embed/instagram' to the
+	// canonical form 'core/embed'.
+	if ( name && name.indexOf( 'core-embed/' ) === 0 ) {
+		// Capture `core-embed/instagram` into `{"providerNameSlug":"instagram"}`
+		const providerSlug = name.substring( 11 );
+		const deprecated = {
+			speaker: 'speaker-deck',
+			polldaddy: 'crowdsignal',
+		};
+		newAttributes.providerNameSlug =
+			providerSlug in deprecated
+				? deprecated[ providerSlug ]
+				: providerSlug;
+		// this is needed as the `responsive` attribute was passed
+		// in a different way before the refactoring to block variations
+		if ( ! [ 'amazon-kindle', 'wordpress' ].includes( providerSlug ) ) {
+			newAttributes.responsive = true;
+		}
+		name = 'core/embed';
+	}
+
+	// Convert 'core/query-loop' blocks in existing content to 'core/post-template'.
+	// TODO: Remove this check when WordPress 5.9 is released.
+	if ( name === 'core/query-loop' ) {
+		name = 'core/post-template';
+	}
+
+	return [ name, newAttributes ];
+}
