+++ packages/blocks/src/api/raw-handling/html-to-blocks.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { createBlock, findTransform } from '../factory';
-import { getBlockAttributes } from '../parser';
+import { getBlockAttributes } from '../parser/get-block-attributes';
 import { getRawTransforms } from './get-raw-transforms';
 
 /**
