+++ packages/blocks/src/api/templates.js
@@ -11,7 +11,7 @@ import { renderToString } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { convertLegacyBlocks } from './parser';
+import { convertLegacyBlockNameAndAttributes } from './parser/convert-legacy-block';
 import { createBlock } from './factory';
 import { getBlockType } from './registration';
 
@@ -108,10 +108,13 @@ export function synchronizeBlocksWithTemplate( blocks = [], template ) {
 				attributes
 			);
 
-			const {
-				name: blockName,
-				attributes: blockAttributes,
-			} = convertLegacyBlocks( name, normalizedAttributes );
+			const [
+				blockName,
+				blockAttributes,
+			] = convertLegacyBlockNameAndAttributes(
+				name,
+				normalizedAttributes
+			);
 
 			return createBlock(
 				blockName,
