+++ test/integration/fixtures/blocks/core__pullquote__deprecated-4.serialized.html
@@ -1,3 +1,3 @@
-<!-- wp:pullquote {"textAlign":"left","className":"has-background has-black-background-color is-style-solid-color","backgroundColor":"black","textColor":"white"} -->
-<figure class="wp-block-pullquote has-text-align-left has-background has-black-background-color is-style-solid-color has-white-color has-black-background-color has-text-color has-background"><blockquote><p>pullquote</p><cite>before block supports</cite></blockquote></figure>
+<!-- wp:pullquote {"textAlign":"left","className":"is-style-solid-color","backgroundColor":"black","textColor":"white"} -->
+<figure class="wp-block-pullquote has-text-align-left is-style-solid-color has-white-color has-black-background-color has-text-color has-background"><blockquote><p>pullquote</p><cite>before block supports</cite></blockquote></figure>
 <!-- /wp:pullquote -->
